+++ demo/addons/godot-openxr/CHANGES.md
@@ -4,3 +4,4 @@ Changes to the Godot OpenXR asset
 1.0.0 - In progress
 -------------------
 - Original implementation
+- Switched to use godot-cpp